requirements.txt
Flask
selenium
requests
pymongo
python-dotenv
bson
